refactor: RenderTabContent 및 RenderCommonContent 컴포넌트 분리 - 상태 관리(setState) 함수의 타입 호환 문제 해결
