+++ src/pages/meetings/index.tsx
@@ -1,15 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import type { NextPage, GetServerSideProps } from 'next';
 import axios from 'axios';
-import Greeting from '@/components/findGatherings/Greeting';
-import EditMeetingButton from '@/components/findGatherings/editMeeting/EditMeetingButton';
-import SubCategoryFilter from '@/components/findGatherings/SubCategoryFilter';
-import PlanFilter from '@/components/findGatherings/PlanFilter';
-import PlanList from '@/components/findGatherings/PlanList';
 import { useCursorInfiniteScroll } from '@/hooks/useCursorInfiniteScroll';
 import { PlanDataWithCategory } from '@/types/plans';
 import { RegionOption, SubRegionOption } from '@/types/reviewType';
 import Tabs from '@/components/findGatherings/tab/Tabs';
+import RenderTabContent from '@/components/findGatherings/RenderTabContent';
 
 const baseUrl = process.env.NEXT_PUBLIC_BASE_URL;
 
@@ -29,72 +25,6 @@ interface HomeProps {
   initialPlans: PlanDataWithCategory[];
   initialCursor: number | null;
 }
-
-// 탭별(달램핏/워케이션) 공통 렌더링 컴포넌트
-// 서브카테고리, 모임 만들기 버튼 (인증필요), 일정카드 목록
-const RenderCommonContent: React.FC<{
-  plans: PlanDataWithCategory[];
-  selectedDate: string | null;
-  setSelectedDate: React.Dispatch<React.SetStateAction<string | null>>;
-  selectedRegion: RegionOption | null;
-  setSelectedRegion: React.Dispatch<React.SetStateAction<RegionOption | null>>;
-  selectedSubRegion: SubRegionOption | null;
-  setSelectedSubRegion: React.Dispatch<
-    React.SetStateAction<SubRegionOption | null>
-  >;
-  selectedCategory: string;
-  selectedSubCategory: string | null;
-}> = ({
-  plans,
-  selectedDate,
-  setSelectedDate,
-  selectedRegion,
-  setSelectedRegion,
-  selectedSubRegion,
-  setSelectedSubRegion,
-  selectedCategory,
-  selectedSubCategory,
-}) => {
-  const [isAuthenticated, setIsAuthenticated] = useState(false);
-
-  useEffect(() => {
-    // 토큰 저장소에 따라 변경 필요
-    const token = localStorage.getItem('accessToken');
-    setIsAuthenticated(!!token);
-  }, []);
-
-  return (
-    <div>
-      {/* PlanFilter: 지역/날짜 선택 필터 */}
-      <PlanFilter
-        selectedDate={selectedDate}
-        setSelectedDate={setSelectedDate}
-        selectedRegion={selectedRegion}
-        selectedSubRegion={selectedSubRegion}
-        onRegionChange={(region) => {
-          setSelectedRegion(region);
-          setSelectedSubRegion(null);
-        }}
-        onSubRegionChange={(sub) => setSelectedSubRegion(sub)}
-      />
-      {/* 모임 만들기 버튼 (인증된 사용자만 노출) */}
-      {isAuthenticated && (
-        <div className="mb-6 flex justify-end">
-          <EditMeetingButton />
-        </div>
-      )}
-      {/* 필터링된 일정 카드 목록 */}
-      <PlanList
-        plans={plans || []}
-        selectedDate={selectedDate}
-        selectedRegion={selectedRegion}
-        selectedSubRegion={selectedSubRegion}
-        selectedCategory={selectedCategory}
-        selectedSubCategory={selectedSubCategory}
-      />
-    </div>
-  );
-};
 
 const Home: NextPage<HomeProps> = ({ initialPlans, initialCursor }) => {
   //상태관리
@@ -144,35 +74,6 @@ const Home: NextPage<HomeProps> = ({ initialPlans, initialCursor }) => {
     setSelectedCategory(activeTab);
   }, [activeTab]);
 
-  // 탭별 콘텐츠 렌더링
-  const renderTabContent = (category: string) => {
-    //setSelectedCategory(category);
-    return (
-      <div className="mx-auto items-center sm:w-[400px] sm:justify-center md:w-[600px] lg:w-full">
-        <Greeting />
-        {/* 달램핏 탭일 경우 하위 카테고리 필터 추가 */}
-        {category === '달램핏' && (
-          <SubCategoryFilter
-            selectedSubCategory={selectedSubCategory}
-            setSelectedSubCategory={setSelectedSubCategory}
-          />
-        )}
-        {/* 공통 콘텐츠 렌더링 */}
-        <RenderCommonContent
-          plans={plans}
-          selectedDate={selectedDate}
-          setSelectedDate={setSelectedDate}
-          selectedRegion={selectedRegion}
-          setSelectedRegion={setSelectedRegion}
-          selectedSubRegion={selectedSubRegion}
-          setSelectedSubRegion={setSelectedSubRegion}
-          selectedCategory={selectedCategory}
-          selectedSubCategory={selectedSubCategory}
-        />
-      </div>
-    );
-  };
-
   return (
     <div className="mx-auto px-4 py-6">
       {/* 탭 컴포넌트 */}
@@ -184,7 +85,22 @@ const Home: NextPage<HomeProps> = ({ initialPlans, initialCursor }) => {
           setCursor(initialCursor);
           setIsFetching(false);
         }}
-        renderContent={renderTabContent}
+        renderContent={(category) => (
+          <RenderTabContent
+            category={category}
+            plans={plans}
+            selectedDate={selectedDate}
+            setSelectedDate={setSelectedDate}
+            selectedRegion={selectedRegion}
+            setSelectedRegion={setSelectedRegion}
+            selectedSubRegion={selectedSubRegion}
+            setSelectedSubRegion={setSelectedSubRegion}
+            selectedCategory={selectedCategory}
+            setSelectedCategory={setSelectedCategory}
+            selectedSubCategory={selectedSubCategory}
+            setSelectedSubCategory={setSelectedSubCategory}
+          />
+        )}
       />
       <div ref={loaderRef} className="h-12"></div>
     </div>
